Changed to integer
git-svn-id: https://svn.code.sf.net/p/comet-ms/code@452 299d2d02-8f3b-485e-a07e-71e8e5e16577

+++ branches/integermath/CometSearch/CometDataInternal.h
@@ -637,8 +637,8 @@ struct Query
 
    // Standard array representation of data
    float *pfSpScoreData;
-   float *pfFastXcorrData;
-   float *pfFastXcorrDataNL;  // pfFastXcorrData with NH3, H2O contributions
+   int *pfFastXcorrData;
+   int *pfFastXcorrDataNL;  // pfFastXcorrData with NH3, H2O contributions
 
    PepMassInfo          _pepMassInfo;
    SpectrumInfoInternal _spectrumInfoInternal;
